requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv

langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langsmith==0.1.147

faiss-cpu==1.7.4
numpy==1.26.4

sentence-transformers
pypdf

openai==1.30.1
httpx==0.27.0